+++ src/ui/view/mod.rs
@@ -5,8 +5,18 @@
 #![doc = crate::_doc!(modules: crate::ui; view)]
 #![doc = crate::_doc!(flat:"ui")]
 #![doc = crate::_doc!(hr)]
+//
+
+// mod cell;
+// mod document;
+// mod graphic;
 
 crate::structural_mods! { // _mods
     _mods {
+        // pub use super::{
+        //     cell::_all::*,
+        //     document::_all::*,
+        //     graphic::_all::*,
+        // };
     }
 }