+++ src/ui/widget/mod.rs
@@ -1,10 +1,10 @@
-// devela/src/ui/intent/mod.rs
+// devela/src/ui/widget/mod.rs
 //
-#![doc = crate::_tags!(wip)]
-#![doc = crate::_DOC_UI_INTENT!()] // public
-#![doc = crate::_doc!(modules: crate::ui; intent)]
+#![doc = crate::_DOC_UI_WIDGET!()] // public
+#![doc = crate::_doc!(modules: crate::ui; widget)]
 #![doc = crate::_doc!(flat:"ui")]
 #![doc = crate::_doc!(hr)]
+//
 
 crate::structural_mods! { // _mods
     _mods {